move array + helper fcn to util file
